backend/agent/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
